debian: Use an shlibs file instead of a symbols file
The necessary symbols in the symbols file differ across architectures,
and are all C++ mangled names which are hard to maintain. Since upstream
have not guaranteed ABI stability yet, it is likely that ABI will break
with each release, so we might as well just use an shlibs file for now.
It will be simpler to maintain until ABI stability is guaranteed.

Signed-off-by: Philip Withnall <withnall@endlessm.com>