debian: Run keygen process as ndn-user
If it’s run as root it will inherit the HOME of root (or the user
running sudo) and hence the keys will probably be generated in ~/.ndn
for the user running sudo.

Instead, the keys need to be in /var/lib/ndn, so these commands must be
run as ndn-user.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T13932